+++ test/runtests.jl
@@ -45,6 +45,7 @@ const SAMPLE_MS = get(ENV, "MEASUREMENTSETS_TEST_MS",
     include("keywords_tests.jl")
     include("endian_tests.jl")
     include("robustness_tests.jl")
+    include("msvalid_tests.jl")
     include("container_tests.jl")
     include("taql_query_tests.jl")
     include("taql_command_tests.jl")